Added feature flag to disable subscription status handling
We want to move to user memberships being what handles if someone is an active member.
The primary reason is that the existing household memberships have subscriptions per member
but that introduces complications and they're $0 subscriptions. The new team/household code
has one subscription (that manages pricing for the group and is attached to one user) while
each user has their own user membership. We did have to prevent user memberships from activating
if no ID check has occurred, but that's handled on the WordPress side.

We also removed the concept of "need id check" and "id was checked" user membership statuses
in favor of a meta data key if the id was checked. This should reduce a lot of logic issues
surrounding membership and ID checks and each piece, ie subscription, user memberhsip, customer
sort of manages what it's good at.

As far as the code goes, in summary:
- Added SUBSCRIPTION_STATUS_IGNORED feature flag to completely disable subscription status changes affecting membership status
- Removed need-id-check and id-was-checked statuses for user memberships in favor of "paused" and "active" being the primary ones