Fixing external membership options lookup
There were 2 issues. The first is the filter was an empty string
which was causing everything to be filtered out. The second was that
after we fixed the filter, the number of options we were returning
was way more than what Slack wanted. We're now limiting the number
of options that are returned.